Fix mnemonics off by one error (#438)
* Simplify mnemonics implementation

* ebin

* Fix build

* Fix off by one error, fix type coercion

+++ src/zedwallet/CommandImplementations.cpp
@@ -263,8 +263,9 @@ void printHashrate(uint64_t difficulty)
     }
 
     /* Hashrate is difficulty divided by block target time */
-    uint64_t hashrate = round(difficulty / 
-                              CryptoNote::parameters::DIFFICULTY_TARGET);
+    uint32_t hashrate = static_cast<uint32_t>(
+        round(difficulty / CryptoNote::parameters::DIFFICULTY_TARGET)
+    );
 
     std::cout << "Network hashrate: "
               << SuccessMsg(Common::get_mining_speed(hashrate))
